@@ -1,6 +1,5 @@
 name: sossoldi
 description: A new Flutter project.
-
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
@@ -20,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.18.0 <3.0.0'
+  sdk: '>=3.1.5 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -37,7 +36,7 @@ dependencies:
   cupertino_icons: ^1.0.2
 
   # Support for various charts
-  fl_chart: ^0.55.2
+  fl_chart: ^0.64.0
 
   # State manager
   flutter_riverpod: ^2.1.3
@@ -53,10 +52,10 @@ dependencies:
   url_launcher: ^6.0.6
 
   # Format DateTime
-  intl: ^0.17.0
+  intl: ^0.18.1
 
   # Display todos in grid
-  flutter_staggered_grid_view: ^0.4.0-nullsafety.3
+  flutter_staggered_grid_view: ^0.7.0
 
   percent_indicator: ^4.2.2
   circular_menu: ^2.0.1
@@ -66,14 +65,14 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_launcher_icons: ^0.12.0
+  flutter_launcher_icons: ^0.13.1
 
   # The "flutter_lints" package below contains a set of recommended lints to
   # encourage good coding practices. The lint set provided by the package is
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^2.0.0
+  flutter_lints: ^3.0.1
 
   # Flutter test
   test: ^1.5.1
