@@ -180,7 +180,7 @@ class TransactionRow extends ConsumerWidget with Functions {
             children: [
               Material(
                 borderRadius: BorderRadius.circular(8),
-                color: Theme.of(context).colorScheme.background,
+                color: Theme.of(context).colorScheme.primaryContainer,
                 child: InkWell(
                   onTap: () {
                     ref.read(selectedTransactionUpdateProvider.notifier).state =
@@ -208,7 +208,7 @@ class TransactionRow extends ConsumerWidget with Functions {
                           decoration: BoxDecoration(
                             shape: BoxShape.circle,
                             color: category?.color != null
-                                ? categoryColorList[category!.color]
+                                ? categoryColorListTheme[category!.color]
                                 : Theme.of(context).colorScheme.secondary,
                           ),
                           child: Padding(
