@@ -66,7 +66,7 @@ class _CategorySelectorState extends ConsumerState<CategorySelector>
                       itemBuilder: (context, i) {
                         CategoryTransaction category = categories[i];
                         IconData? icon = iconList[category.symbol];
-                        Color? color = categoryColorList[category.color];
+                        Color? color = categoryColorListTheme[category.color];
                         return Padding(
                           padding: const EdgeInsets.symmetric(horizontal: 16.0),
                           child: Column(
@@ -130,7 +130,7 @@ class _CategorySelectorState extends ConsumerState<CategorySelector>
                     itemBuilder: (context, i) {
                       CategoryTransaction category = categories[i];
                       IconData? icon = iconList[category.symbol];
-                      Color? color = categoryColorList[category.color];
+                      Color? color = categoryColorListTheme[category.color];
                       return ListTile(
                         tileColor: Theme.of(context).colorScheme.surface,
                         onTap: () => ref.read(categoryProvider.notifier).state =
