@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../constants/style.dart';
+import 'add_page/add_page.dart';
 import 'home_page.dart';
 import 'planning_page/planning_page.dart';
 import 'statistics_page.dart';
@@ -39,22 +40,17 @@ class _StructureState extends ConsumerState<Structure> {
   Widget build(BuildContext context) {
     final selectedIndex = ref.watch(selectedIndexProvider);
     return Scaffold(
-      backgroundColor: blue7,
-      // Prevent the fab moving up when the keyboard is opened
-      resizeToAvoidBottomInset: false,
+      // backgroundColor: blue7,
+      resizeToAvoidBottomInset:
+          false, // Prevent the fab moving up when the keyboard is opened
       appBar: AppBar(
         // Sulla dashboard (0) setto il background blue
-        backgroundColor: selectedIndex == 0
-            ? blue7
-            : Theme.of(context).colorScheme.background,
+        // backgroundColor: selectedIndex == 0 ? blue7 : Theme.of(context).colorScheme.background,
         elevation: 0,
         centerTitle: true,
         title: Text(
           _pagesTitle.elementAt(selectedIndex),
-          style: Theme.of(context)
-              .textTheme
-              .headlineLarge!
-              .copyWith(color: Theme.of(context).colorScheme.primary),
+          style: Theme.of(context).textTheme.headlineLarge!,
         ),
         leading: Padding(
           padding: const EdgeInsets.only(left: 16),
@@ -97,10 +93,10 @@ class _StructureState extends ConsumerState<Structure> {
       bottomNavigationBar: BottomNavigationBar(
         type: BottomNavigationBarType.fixed,
         selectedItemColor: Theme.of(context).colorScheme.primary,
-        unselectedItemColor: grey1,
+        // unselectedItemColor: grey1,
         selectedFontSize: 8,
         unselectedFontSize: 8,
-        backgroundColor: const Color(0xFFF6F6F6),
+        // backgroundColor: const Color(0xFFF6F6F6),
         currentIndex: selectedIndex,
         onTap: (index) => index != 2
             ? ref.read(selectedIndexProvider.notifier).state = index
